requirements.txt
bcrypt==4.0.1
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
dnspython==2.4.2
Flask==2.3.2
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
marko==2.1.2
MarkupSafe==2.1.5
packaging==24.1
pymongo==4.4.1
python-dotenv==1.0.0
Werkzeug==3.0.3
